8.04g : update after debug/bug fixing
When logging to file the emonCMS updates take at least twice as long
as when SD card writes are not occurring - so a high load due to SD.
Therefore, don't have simultaneous debug file logging and emon updates.

Fix the start time for UserIO so uptime is generated correctly.

Use PW_HP_MODBUS to log LG modbus data traffic (RTU only) - data is
captured to SD in /hpmodbus.log.  This is sent via email on boot and
is deleted following the send.

Quick debug on sends/fails to emon.

When sending to emon free bytes drops from ~140 KiB to 80 and back.
There isn't a great deal of DDR in any event (probably 320 KiB is the
DRAM available with a trivial image).