8.06c: Re-use emon connection to reduce core 0 usage, new comms screen
Problem with emon comms is not keeping the connection open to the server
when a number of quick feed updates are required - so the TLS exchange
takes time for each feed update.  Re-use the connection if possible.

New screen to keep track of emon send requests/background Q failures and
emon feed rejections.

LG modbus update - write to the registers.log at 128 intervals so
we can determine how far through the map the processing has progressed.