9.8.0a - working flow rate
Added the quadratic parameters for all UPS3 modes and tested.  If
UPS3 power is less than mode minimum, zero flow rate is returned.
If above the mode maximum then flow rate error returned - when the
UPS3 is at the maximum power level for the mode then it can deliver
flow but we can't calculate it as the power is at max.

Next step is to add to emon..