10.0.8b : Initial changes for monitor board
If pin 15 is detected low then we set RxD, TxD to GPIO's 32 and 33
to ensure any stray Serial.print's won't be generated.

For a monitor board the intention is use the default Rx/Tx pins for
the modbus communication.  So the monitor will have a switch to pull
pin 15 to ground.  This also stops the ESP32 from output any serial
output on power cycle (supposedly).

GPIO for PMW is no longer from the sensors.dat but from the hwconfig.h,
as per all the other ESP32 related pin assignments (could ultimately
place those in a config file in SPIFFS).

All use of TEMPERATURE_BOARD to select functionality has been replaced
with checks for not being as MASTER_BOARD.  In fact most of these were
for SD detection so really should change to that feature availability.

So changes required will be for utils.cpp to stop any serial output
via the sketch debug macros if pin 15 is low - currently serial output
is being generated on pins 32 and 33 theoretically.

And then onto more debugging & analysis of the monitor board