10.0.8d - h/w detect serial enabled
Standard ESP32 behaviour is that if GPIO 15 is low at boot then there
is no serial output during boot.  We extend that to prevent any serial
debug output from our code, and as the Serial0 rx/tx pins are also
remapped then other library use of Serial will not be on the default
Serial-0 IO pins.

Use internal pull-up on GPIO-15 that should be ok for the all boards,
with the monitor board having a switch to ground to disable serial-0
and thereby map for ModBus.

The modbus changes are not confirmed yet - we need to disable modbus
if serial port is active.