10.0.8e - Updates for the monitor board
If modbus isn't available then do not instantiate a heat pump - this
can happen when a monitor board has serial debug enabled as the same
h/w UART is used for serial and modbus comms.

Allow touch events for any board type if set in the build time configs.

Monitor using GPIO13 for one-wire and GPIO12 for modbus enable as we
need to pull up one-wire to Vcc and GPIO12 has to be low for boot-up.

Everything tested ok on monitor with breadboard for IO - except PMW
not tested.  Will have to just assume all ok.

Needs testing on master board to ensure nothing broken.