12.0.3 - support for external board, minor changes for heating temps
UserIO changes to replace 'rx' with 'Err' for modbus errors.

New hardware config for the external board, uses GPIO13 for one-wire,
no PWM support.

UFH temperatures references changed to 'heating'.

Added the firmware update batch file so Arduino not required for
programming.  A default spiffs can be created so the s/w doesn't have
to format the SPIFFS (it will do on error).