12.0.4 - reset replaces .dat files with .def files
Noted that large files couldn't be edited on a t-node.

Outstanding what to do with the LG mapping.  We could have another
extension, '.hid' that are not displayed (unless by config) - then
if LG is enabled if sensors.dat and lg.dat is not available then the
.hid is copied to a .def, which is then copied to a .dat ?

And if LG not in sensors and .dat exists then delete it ?