Webserver using HTTP_POST for editing files - hopefully fix lg reg. crash ?
Previously used GET with the URL formed as the ?edit_textarea=<text>,
so the length of the URL is approx. the length of the text changed.

Now use a POST with the edit_textarea as a parameter which is then
handled in the body of the HTTP message and not a URL.  This allows
other browsers to edit files > 2 KB (which is the limit of URL's really,
chrome allow larger URLs).

With this can edit text files up to 18 KiB successfully, but if the
text file contains '%' then that may screw up the webserver processor.
Also a 19 KiB file resulted in zero bytes being returned, so another
bug in the async webserver/tcp code.  And at 20 KiB would have an
exception causing crash.

Should really put a 16 KiB limit on the file to be edited.