Fixed large file download issue
The async. web server library would not download files > ~ 5 KiB, it
appeared to be due to trying to transfer more than 5300 bytes in the
first chunked transfer.  Limit the chunk length to 4 KiB resolved
the issue.

Next is to move some code to the WebServer class and out of global
scope.