12.0.7 - Large networking update
The problem with saving LG sensor data was due to lack of heap space.
With the HttpClient used by emon taking ~ 50KB of heap, then the change
set here is to delete the HttpClient when editing a file.  Then have
to timeout this file editing (2 minutes) and restart the ESP if not
returned back to the homescreen.

So quite a few changes to accomodate that, and tidied up some of the
existing code surrounding the networking (web client) mutex used by
email, modbus TCP and emoncms updating.

With this change set could edit larger files and expect that this will
be ok on the master monitor & LG sensor edits.  May still be better to
download/upload as opposed to in-situ editing.

(There is another change required to the IP designator for email which
is currently hard coded to a domain that needs to be taken from the
SMTP host or account name really).