MAJOR UPDATE: Arduino IDE to 2.3.4 and esp32 to 3.1.0
Looking at using a new Waveshare esp32 module and this is easier if the
underlying ESP release from Expressif is used. Unfortunately this had a
huge impact in that the previous s/w would no longer fit in the existing
flash space.  So a repartitioning exercise is required for all boards to
allow the later IDE and Expressif core to be used.

So now we have 512 MiB SPIFFS and 2x OTA apps of 1728 KiB.  But existing
boards have to be re-flashed with their data files from existing SPIFFS
re-populating the smaller FS partition.

This comes on the back of some other major changes reqarding networking
that hopefully resolves large file editing that is hampered by the
memory use of HttpClient.

External libraries have also been updated as the Expressif changes
introduced an assert in the lwip TCP stack with async. web server.

Pulled the git repo's for the esptool python and esp-idf which are then
used by the new facreset.sh script.  This will need some further mods but
is a more advanced version than the previous DOS based solution.  Now
vanilla 4MiB images can be created to completely reset flash status.