v25.1.6 - improve user experience on reset
When a board is reset then marker file is written and on reboot the
UserIO displays that in AP mode due to reset.  The reset marker is only
removed when the reboot from webserver happens (or user deletes the
marker file - which should really be hidden).

Upon reboot the lack of reset marker results in the boot proceeding as
normal.

So sequence for commissioning a board:-

   flash it, default & current data files in spiffs
   make any modifications/test etc
   reset the board via server - marker is written, .def -> .dat occurs
   board starts in AP mode and is ready for consumer
      test AP mode active and .dat files are as expected