v25.01.11 - rework temperatures, reset handling, config at boot
Overnight master monitor lost temperature data.  Values of -127 being
reported for the HP return, but outside actually matched the HP
return.  Not clear even how a value of -127 can propagate to emon.  The
system did recover, without reboot - but never seen this before.

As part of work to receover memory then a new Sample was not copied
from the current sample, perhaps that is causing a problem ?  Not had
enough time to debug the problem.  So, added increased mutex protection
around the temperatures - with the view that perhaps async UDP updates
caused problems.  Removed copies of temperature sensors so only using
existing pointers.  As we don't create new sensors then these are
fixed effectively.

In reset handling now use the esp_reset_reason() to get the reboot
code and map it to our RebootType.  A map of RebootType to String is
provided for convenience - all within the Config class. Provisional
function in place to detect fast reboots, but not implemented.

The serial output at boot has been compromised - again there was
recursion occuring. So stopped that recursive behaviour - early PW_
messages will be sent to Serial (if boot serial enabled) until the
Config instance is available which will then be used to filter messages.