v25.01.15 - fix temperatures recorded when device is lost off the 1-Wire bus
Recently the HP Return temperature has been sporadic and seen that the
device is lost from the bus, but then temperatures are being swapped
between HP flow and outside for example.

The likely reason is that when a device is lost from the bus that the
original bus index for DS18B20's are then no longer valid.  So temps
are recorded against the wrong device.

Possible fix by using the device address as opposed to its original
bus index.  This appeared to work with local testing, with hard removal
of a device from the bus.