25.08.01 - Defensive changes to prevent bad data to emon
Previously seen bad data sent to emon, especially from the LG, if data
is not read correctly.  In the outtage this month (due to electrical
reworking the pump contactors) then this was very apparent.

Don't send DS temperature data to emon if > 60deg.  Seen the DS read
ok, but bad data (128 deg).  Therefore this may help prevent bad emon
data.

For LG, if modbus error is detected in a sample then do not send any
data to emon by effectively making the LG registers null pointers in
the current sample.