Added new board to arduiono esp32 boards.txt : pw_wshare
This is a copy of the WAVESHARE_ESP32_S3_TOUCH_LCD_43B device but with
the addition of changes to build.defines.

   pw_wshare.build.defines={build.psram_defines} -DPW_LCD=1

   pw_wshare.menu.PSRAM.enabled.build.psram_defines=-DBOARD_HAS_PSRAM

Normally the latter is not psram_defines but simply defines.  The issue
is that build.defines is not hierarchic - so if any later menu/option
changes the build.defines then it sets a new value. So we use the
method here of assigning build defines via subitems.

The PW_LCD is used (needs redefinition).