Version string moved from header, identified remote therm. problem
If UDP logging disabled at boot then the call to AsyncUDP::listen()
succeeds as it creates a UDP task.  If the UDP logging is enabled at
boot then a call to connect() will be made that also creates the task.
A subsequent call to listen() after connect() will fail.  So need to
patch listen() to continue if the task already exists.

Also need to have more than 2K stack for the async_udp task otherwise
stack depth exceeded when UDP packets received.

Not fully tested yet but most likely the solution to a long standing
problem.