Fix reset processing.
No longer use a file presence in SPIFFS but the persistent reboot
reason that will be set to SERVER_RESET.  This then falls back to the
usual new configuration processing and on reboot via the server then
the board will boot to normal mode.  Note that power cycling will not
reset the reboot code, a server reboot is required.