Send stored data at same daily update time.  Improve Ardiono boards management.
Stored data is sent at daily update time, not at midnight.

The boards.txt in the esp32 Arduino library has been reverted to the
original, and custom boards added to a new file boards.local.txt that
Arduino IDE picks up.  Removed the board variant as it is an esp32
and simplified menu structure as hard coded default options.