Switch to ReadyMail email library & used fixed event line length in LG log
In order to send html as body text in an email as opposed to simply
sending an attachment we switched to the ReadyMail libary.  This has
some patches added to improve efficiency and fix an issue when only
sending body text from html file.  We add 2 new defines in build opts

  READYMAIL_OPTIMISED=1                # to use optimisation changes
  READYMAIL_EXTERNAL_BUFF_SIZE=4064    # for ReadyMail to use the scratch
                                       # buffer

So the LG event log is now sent as a body as opposed to attachment.  The
SD current data is still sent as an attachment.

Email file attachment/body now default to SPIFFS not SD.

Add a loopTest() in ThermaV for calling from 'Run' webserver button.

Events added to the LG event log are now set at 300 characters using
HTML comment block to pad out the data.  This is to facilitate the
removal of old events from the log.  So the intention is to eventually
fix the file size to N events and then remove old events, this change
will improve the ability to seek to the N'th event to remove old lines
perhaps.